split more
